+++ examples/scripts/compare_lead_acid.py
@@ -9,7 +9,7 @@ pybamm.set_logging_level("INFO")
 models = [
     pybamm.lead_acid.LOQS(),
     pybamm.lead_acid.FOQS(),
-    pybamm.lead_acid.CompositeExtended(),
+    pybamm.lead_acid.Composite(),
     pybamm.lead_acid.Full(),
 ]