+++ tests/unit/test_expression_tree/test_operations/test_copy.py
@@ -29,7 +29,6 @@ class TestCopy(unittest.TestCase):
             pybamm.FunctionParameter("function", {"a": a}),
             pybamm.grad(v_n),
             pybamm.div(pybamm.grad(v_n)),
-            pybamm.Integral(a, pybamm.t),
             pybamm.IndefiniteIntegral(v_n, x_n),
             pybamm.BackwardIndefiniteIntegral(v_n, x_n),
             pybamm.BoundaryValue(v_n, "right"),