+++ tests/unit/test_expression_tree/test_operations/test_simplify.py
@@ -72,8 +72,19 @@ class TestSimplify(unittest.TestCase):
 
         # Integral
         self.assertIsInstance(
-            (pybamm.Integral(a, pybamm.t)).simplify(), pybamm.Integral
+            (
+                pybamm.Integral(a, pybamm.SpatialVariable("x", domain="domain"))
+            ).simplify(),
+            pybamm.Integral,
         )
+
+        def_int = (pybamm.DefiniteIntegralVector(a, vector_type="column")).simplify()
+        self.assertIsInstance(def_int, pybamm.DefiniteIntegralVector)
+        self.assertEqual(def_int.vector_type, "column")
+
+        bound_int = (pybamm.BoundaryIntegral(a, region="negative tab")).simplify()
+        self.assertIsInstance(bound_int, pybamm.BoundaryIntegral)
+        self.assertEqual(bound_int.region, "negative tab")
 
         # BoundaryValue
         v_neg = pybamm.Variable("v", domain=["negative electrode"])