+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-# Unreleased
+# [Unreleased](https://github.com/pybamm-team/PyBaMM)
 
 ## Features
+
+-   Add averaging in secondary dimensions ([#1057](https://github.com/pybamm-team/PyBaMM/pull/1057))
 
 ## Optimizations